refactor: Move TargetingBuilder and siblings out of shared folder
Initially I thought I might implement simpler versions of workflow targeting and email campaign targeting
They would share the targeting functionality.

I ended up creating a deeper implementation of Workflow instead and focusing on it.

So at this point none of that is share and it makes much more sense this way