Rename to email campaign and segmentation
When I started this project, I didn't have a great idea of the terminology that would be used. This PR refactors the key terms in the code to more appropriate ones.

This is a testament to the power of the test suite I've built up. I made a dozen commits touching dozens of files at a time, and  was able to move quickly and confidently as long as the tests were passing.